AI_Weed_Detection/Backend/requirements.txt
fastapi
uvicorn
torch
torchvision
ultralytics
opencv-python
pillow
numpy<2
python-multipart
